move is_retryable_error in utils
